new lib for getReleaseReportReferenceMessage
