+++ lib/chat/toolChains/create_release_report/createReleaseReportToolChain.ts
@@ -1,23 +1,5 @@
 import { ToolChainItem } from "../toolChains";
-import { ModelMessage } from "ai";
-import { referenceReleaseReport } from "./referenceReleaseReport";
-
-/**
- * Creates a reference message with the release report example
- */
-const getReleaseReportReferenceMessage = (): ModelMessage => {
-  return {
-    role: "user",
-    content: [
-      {
-        type: "text" as const,
-        text: `Here is an example release report for reference. Use this as a template for creating your own release reports:
-
-        ${referenceReleaseReport}`,
-      },
-    ],
-  };
-};
+import getReleaseReportReferenceMessage from "./getReleaseReportReferenceMessage";
 
 export const createReleaseReportToolChain: ToolChainItem[] = [
   { toolName: "search_web", messages: [getReleaseReportReferenceMessage()] },
